Split files vuex into multiple files for better maintainability
