+++ apps/files/src/store/actions.js
@@ -1,29 +1,4 @@
-import { findIndex, without, map, filter } from 'lodash'
-import { fileFilters } from './fileFilters.js'
-const namespaced = true
-
-const state = {
-  currentFolder: null,
-  files: [],
-  filesSearched: [],
-  fileFilter: fileFilters,
-  selected: [],
-  inProgress: [],
-  searchTermGlobal: '',
-  searchTermFilter: '',
-  davProperties: [
-    '{http://owncloud.org/ns}permissions',
-    '{http://owncloud.org/ns}favorite',
-    '{http://owncloud.org/ns}fileid',
-    '{http://owncloud.org/ns}owner-id',
-    '{http://owncloud.org/ns}owner-display-name',
-    '{DAV:}getcontentlength',
-    '{http://owncloud.org/ns}size',
-    '{DAV:}getlastmodified',
-    '{DAV:}getetag',
-    '{DAV:}resourcetype'
-  ]
-}
+import map from 'lodash'
 
 function _buildFile (file) {
   let ext = false
@@ -77,84 +52,7 @@ function _buildFile (file) {
   })
 }
 
-const mutations = {
-  UPDATE_FILE_PROGRESS (state, progress) {
-    let fileIndex = findIndex(state.inProgress, (f) => {
-      return f.name === progress.fileName
-    })
-    if (fileIndex === -1) return
-    state.inProgress[fileIndex].progress = progress.progress
-  },
-  REMOVE_FILE_FROM_PROGRESS (state, file) {
-    let fileIndex = findIndex(state.inProgress, (f) => {
-      return f.name === file.name
-    })
-    state.inProgress.splice(fileIndex - 1, 1)
-  },
-  ADD_FILE_TO_PROGRESS (state, file) {
-    state.inProgress.push({
-      id: file.id,
-      name: file.name,
-      type: file.type,
-      size: file.size,
-      progress: 0,
-      action: 'upload'
-    })
-  },
-  LOAD_FILES (state, { currentFolder, files }) {
-    state.currentFolder = currentFolder
-    state.files = files
-  },
-  LOAD_FILES_SEARCHED (state, files) {
-    state.filesSearched = files
-  },
-  ADD_FILE_SELECTION (state, file) {
-    state.selected.push(file)
-  },
-  REMOVE_FILE_SELECTION (state, file) {
-    if (state.selected.length > 1) {
-      state.selected = without(state.selected, file)
-      return
-    }
-    state.selected = []
-  },
-  RESET_SELECTION (state) {
-    state.selected = []
-  },
-  FAVORITE_FILE (state, item) {
-    let fileIndex = findIndex(state.files, (f) => {
-      return f.name === item.name
-    })
-    state.files[fileIndex].starred = !item.starred
-  },
-  ADD_FILE (state, file) {
-    state.files.push(file)
-  },
-  REMOVE_FILE (state, file) {
-    state.files = without(state.files, file)
-  },
-  SET_SEARCH_TERM (state, searchTerm) {
-    state.searchTermGlobal = searchTerm
-  },
-  SET_FILTER_TERM (state, filterTerm) {
-    state.searchTermFilter = filterTerm
-  },
-  SET_FILE_FILTER (state, filter) {
-    let i = findIndex(state.fileFilter, (f) => {
-      return f.name === filter.name
-    })
-    state.fileFilter[i].value = filter.value
-  },
-  RENAME_FILE (state, { file, newValue, newPath }) {
-    let fileIndex = findIndex(state.files, (f) => {
-      return f.name === file.name
-    })
-    state.files[fileIndex].name = newValue
-    state.files[fileIndex].path = '/' + newPath + newValue
-  }
-}
-
-const actions = {
+export default {
   updateFileProgress ({ commit }, progress) {
     if (progress.progress === 100) commit('REMOVE_FILE_FROM_PROGRESS', { name: progress.fileName })
     else commit('UPDATE_FILE_PROGRESS', progress)
@@ -249,64 +147,4 @@ const actions = {
       })
     })
   }
-}
-
-const getters = {
-  inProgress: state => {
-    return state.inProgress
-  },
-  selectedFiles: state => {
-    if (state.selected.length === 0) {
-      return []
-    } else {
-      return state.selected
-    }
-  },
-  files: state => {
-    return state.files
-  },
-  currentFolder: state => {
-    return state.currentFolder
-  },
-  filterTerm: state => {
-    return state.searchTermFilter
-  },
-  searchTerm: state => {
-    return state.searchTermGlobal
-  },
-  atSearchPage: state => {
-    return state.searchTermGlobal !== ''
-  },
-  fileFilter: state => {
-    return state.fileFilter
-  },
-  activeFiles: state => {
-    // if searchTermGlobal is set, replace current file list with search results
-    let files = state.searchTermGlobal ? state.filesSearched : state.files
-    // respect file filters set in TopBar
-    return filter(files, (file) => {
-      for (let filter of state.fileFilter) {
-        if (file.type === filter.tag) {
-          if (!filter.value) return false
-        } else if (file.name.startsWith('.')) {
-          // show hidden files ?
-          if (!state.fileFilter[2].value) return false
-        }
-      }
-      // respect filename filter for local 'search' in open folder
-      if (state.searchTermFilter && !file.name.toLowerCase().includes(state.searchTermFilter.toLowerCase())) return false
-      return true
-    })
-  },
-  davProperties: state => {
-    return state.davProperties
-  }
-}
-
-export default {
-  namespaced,
-  state,
-  actions,
-  mutations,
-  getters
 }
